GetAccelerationStructureDeviceAddress() added for Vulkan conformance. InstanceData structure size is now 64. CopyImage() takes src and dst imageLayout as parameters. GetShaderGroupHandles() and enum class ShaderGroupHandleType added to Pipeline(), and ShaderBindingTable.cpp/.h added crossplatform only! MIRU_SHADER_COMPILER better handling of input attachment building spv. MIRU_TEST updated. RT is working in both D3D12 and Vulkan in Release; still WIP. D3D12 better handling of instance data for TLAS. d3d12::BindDescriptorSets(), d3d12::BuildAccelerationStructure(), d3d12::~Context() and d3d12::~DescriptorSet() updated. Vulkan better handling of null StridedDeviceAddressRegion and build separate BLASs and TLASs. vulkan DescriptorSet::AddAccelerationStructure actually pushes back to the write descriptor vector.

+++ MIRU_CORE/src/miru_core_common.h
@@ -83,9 +83,9 @@ namespace miru
 		return b;
 	};
 
-	auto alignedSize = [](uint32_t value, uint32_t alignment) -> uint32_t
+	auto alignedSize = [](uint32_t value, uint32_t alignment) -> size_t
 	{
-		return (value + alignment - 1) & ~(alignment - 1);
+		return static_cast<size_t>((value + alignment - 1) & ~(alignment - 1));
 	};
 }
